src/main/java/kz/home/ys/algo/medium/removeDuplicatesFromSortedListII/RemoveDuplicatesFromSortedListII.java
package kz.home.ys.algo.medium.removeDuplicatesFromSortedListII;

class RemoveDuplicatesFromSortedListII {
    public ListNode deleteDuplicates(ListNode head) {
        if (head == null) return null;
        
        ListNode dummy = new ListNode(Integer.MIN_VALUE, head);
        ListNode prev = dummy;
        ListNode curr = head;
        while (curr != null) {
            while (curr != null && curr.next != null && curr.val == curr.next.val) {
                curr = curr.next;
            }
            if (prev.next == curr) {
                prev = prev.next;
                curr = curr.next;
            } else {
                prev.next = curr.next;
                curr = curr.next;
            }
        }
        return dummy.next;
    }
}